Add 'script' category and views/implementation for it

+++ srb2-mod-management/ViewModels/Components/ReleaseViewModel.cs
@@ -210,20 +210,27 @@ namespace srb2_mod_management.ViewModels.Components
 
             foreach (var filepath in downloaded)
             {
-                // For parted files, only extract the first file
-                if (new[]{".zip", ".rar"}.Any(ext => filepath.EndsWith(ext)) 
-                    && !Regex.Match(filepath, @"(part[_ +-.]?(?:0?[2-9]|1\d])|7z.\d?\d[1-9])", RegexOptions.IgnoreCase).Success)
+                if (new[] {".wad", ".lua"}.Any(ext => filepath.EndsWith(ext)))
+                    extractedFiles.Add(filepath);
+
+                else
                 {
-                    using (var archive = ArchiveFactory.Open(filepath))
-                        foreach (var entry in archive.Entries)
-                            if (!entry.IsDirectory)
-                            {
-                                extractedFiles.Add(entry.Key);
-                                entry.WriteToDirectory(path,
-                                    new ExtractionOptions {ExtractFullPath = true, Overwrite = true});
-                            }
+                    if (new[] {".zip", ".rar"}.Any(ext => filepath.EndsWith(ext))
+                        && !Regex.Match(filepath, @"(part[_ +-.]?(?:0?[2-9]|1\d])|7z.\d?\d[1-9])",
+                            RegexOptions.IgnoreCase).Success)
+                    {
+                        using (var archive = ArchiveFactory.Open(filepath))
+                            foreach (var entry in archive.Entries)
+                                if (!entry.IsDirectory)
+                                {
+                                    extractedFiles.Add(entry.Key);
+                                    entry.WriteToDirectory(path,
+                                        new ExtractionOptions {ExtractFullPath = true, Overwrite = true});
+                                }
+                    }
+
+                    File.Delete(filepath);
                 }
-                File.Delete(filepath);
             }
 
             // Add to repository
